cert-fix: avoid crash on missing CS.cfg param
`pki-server cert-fix` reads (and writes) the CS.cfg parameter
`selftests.container.order.startup`.  If this parameter is missing,
the resulting `KeyError` crashes the program.  We have seen several
cases where this parameter is missing, and its absense is otherwise
benign.

Update the relevant subroutine to avoid a crash in the case where
the `selftests.container.order.startup` parameter is missing.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1930586