Bug1971732-subCA Installation cert validation failure
This patch fixes the issue where the CA signing cert is not imported
properly into the nssdb with trust.
The pki cli command is changed from 'nss-import-cert' to 'client-import-cert'
and '--cert' changed to '--ca-cert'.
See https://github.com/dogtagpki/pki/wiki/PKI-Client-CLI#importing-ca-certificate
In addition, if pkispawn fails the pki-server subsystem-cert-validate call,
it will provide more detail on the failure while allow pkispawn to complete.

This would allow admins to manually add the ca signing cert manually.
(Although with the fix mentioned above, it should not be encountered)

https://bugzilla.redhat.com/show_bug.cgi?id=1971732