Drop dependency on xerces-j2 & xalan-j2 to prepare for Java 17
I tested our code against the latest Java 17 as at some point Fedora
will start considering using Java 17 LTS version as the default.

Everything looks good except for some packages brought in by xerces-j2
that are also provided by the JRE. Newer versions of Java are increasing
the encapsulation of the JRE/stricter on name clashes with JRE
internals.

It transpires that the classes we need from xerces-j2 are the very ones
provided by the JRE, so we can simply drop xerces-j2 as a dependency.

As the classes are already present in the Java 11 JRE it is safe to drop
now, so I have done so.

As we only require xalan-j2 to satisfy dependencies of xerces-j2, we can
also drop this as a dependency.