Bug-2008319-pkispawn EC subCA 2-step failure in FIPS mode
This patch addressed issue reported on failure with EC subCA 2-step
installation in FIPS mode.
My investigation reveals that certutil -R would fail in create_request nssdb.py.
I found two different commits having attempts to fix the same issue:

    f1f32c3
    8bf682a
However, the the resulting code appears to NOT contain the working copy.

I pulled out code in this area in 10.2 as a reference (carefully referenced)
to address the issue. It appears to fix the problem. I was able to create
subCA successfully in the FIPS environment on both soft token and hsm.

Note: The pkispawn parameter "pki_req_ski=DEFAULT" introduced in the two
contradicting commits mentioned above doesn't seem to work with the
source right out of the tip of DOGTAG_10_5_BRANCH;  I tried to touch as
little as possible and not alter the logic involving pki_req_ski, and
the result remains the same error as before this patch.

fixes https://bugzilla.redhat.com/show_bug.cgi?id=2008319