Fix NSSExtensionGenerator.createSANExtension()
Previously the NSSExtensionGenerator.createSANExtension() would
add a SAN extension to all certs or requests created with it.

The code has been modified to add a SAN extension only if the
subjectAltName param is specified. The param currently supports
two options:

- DNS:request_subject_cn which will copy the DNS name from the
  CN attribute of the request's subject name
- DNS:request_san_ext which will copy the DNS names from the
  request's SAN extension

The sslserver.conf has been modified to use the new param.