Remove pki_client_database_purge from examples
Previously the pki_client_database_purge param was used in
the examples to retain the temporary NSS database used to
generate the admin certificate so it can be reused after
installation. The param has been removed from the examples
since it's recommended to import the admin PKCS12 file
instead of reusing the temporary NSS database.