Update test for external CA installation
The test for CA installation with external signing cert has
been updated to add a Microsoft Subordinate CA extension in
the CSR and in the issued cert.

The pkispawn default configuration file has been modified to
no longer include the extension since it's already provided
as an example in the subca_signing.conf.