Fix pkidestroy for KRA with external certs
Generally, when installing KRA pkispawn will create a subsystem
cert in the CA using a profile that will also create a subsystem
user for the KRA in the CA (see SubsystemGroupUpdater), then
pkispawn will create a KRA connector in the CA as well (see
UpdateConnector). In case the KRA is removed later pkidestroy can
use this user to remove the KRA connector from the CA.

If the KRA was installed with external certs (including CMC case),
pkispawn would create the subsystem cert using a different profile
which would not create the subsystem user, but it would still
create the KRA connector. However, that means pkidestroy would not
be able to remove the KRA connector.

To fix the problem, some of the code in SubsystemGroupUpdater has
been copied into UpdateConnector such that the subsystem user will
be created when the KRA connector is added if it does not already
exist.

Right now the SubsystemGroupUpdater might still be needed for
other things, but potentially it could be removed in the future.