Fix pkidestroy for OCSP with external certs
Generally, when installing OCSP pkispawn will create a subsystem
cert in the CA using a profile that will also create a subsystem
user for the OCSP in the CA (see SubsystemGroupUpdater), then
pkispawn will create an OCSP publisher in the CA as well (see
UpdateOCSPConfig). In case the OCSP is removed later pkidestroy
can use this user to remove the OCSP publisher from the CA.

If the OCSP was installed with external certs (including CMC
case), pkispawn would create the subsystem cert using a different
profile which would not create the subsystem user, but it would
still create the OCSP publisher. However, that means pkidestroy
would not be able to remove the OCSP publisher.

To fix the problem, some of the code in SubsystemGroupUpdater has
been copied into UpdateOCSPConfig such that the subsystem user
will be created when the OCSP publisher is added if it does not
already exist.

Right now the SubsystemGroupUpdater might still be needed for
other things, but potentially it could be removed in the future.