Redirect banner prompt to standard error
Previously if PKI server was configured with access banner pki
CLI would display the access banner on the standard output.
This could cause a problem if pki CLI is used in automation
because the output might change.

To fix the problem pki CLI has been modified to dispaly the
access banner on the standard error instead.