Set deactivateDefaultTyping() on JSONSerializer ObjectMapper objects
With this setting activated it is possible for bad actors to create
objects not of the type clazz - e.g. a class that extends clazz - which
could introduce malicious code into the application. Explicitly
disabling it guarantees that only an object of type clazz can be
produced.

The remaining ObjectMapper objects in the code should me migrated to use
this class.

Also throw the more specific JsonProcessingException instead of
Exception.