Replace %package_option macro
Previously the %package_option macro was used to include
packages into the build or exclude them from the build
depending on whether the build.sh was called using
--with-pkgs or --without-pkgs param. The problem is the
code has become complicated and difficult to maintain.

To simplify the code the macro has been replaced with a
simple %bcond_without, then the build.sh will convert it
into %bcond_with if the package is excluded.