cli: handle t.getCause() == null in exception handler
It is possible that the cause of a ProcessingException is not set.
In that case, t.getCause() returns null.  As a consequence,
handleException() throws an unhandled NullPointerException.

Check that the cause Throwable is non-null before "drilling down" to
it.

Signed-off-by: Fraser Tweedale <ftweedal@redhat.com>
(cherry picked from commit 70496404fb32063c76d5c7f545b12fbe9ed45ce4)