Add offline support for AuthInfoAccessExtDefault
Previously the AuthInfoAccessExtDefault could only be used inside
the server since it got the port number for the OCSP URL from the
engine object, which obtained the port from the config files (see
CMSEngine.configurePorts()) on startup.

To improve its usability the code has been modified to get the
port number directly from the config files if it's running outside
of the server.